refactored the addtodo and addproject event listeners
